+++ lib/govuk_schemas.rb
@@ -5,8 +5,13 @@ require "govuk_schemas/document_types"
 require "govuk_schemas/example"
 
 module GovukSchemas
-  # @private
-  CONTENT_SCHEMA_DIR = ENV["GOVUK_CONTENT_SCHEMAS_PATH"] || "../publishing-api/content_schemas"
+  def self.content_schema_dir=(path_to_schemas)
+    @content_schema_dir = path_to_schemas
+  end
+
+  def self.content_schema_dir
+    @content_schema_dir ||= ENV.fetch("GOVUK_CONTENT_SCHEMAS_PATH", "../publishing-api/content_schemas")
+  end
 
   # @private
   class InvalidContentGenerated < RuntimeError