+++ lib/govuk_schemas/document_types.rb
@@ -3,7 +3,7 @@ module GovukSchemas
     # Return all of the document types on GOV.UK
     def self.valid_document_types
       @valid_document_types ||= begin
-        filename = "#{GovukSchemas::CONTENT_SCHEMA_DIR}/allowed_document_types.yml"
+        filename = "#{GovukSchemas.content_schema_dir}/allowed_document_types.yml"
         YAML.load_file(filename)
       end
     end