+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# UNRELEASED
+
+* Update default content schemas url to point to publishing api rather than govuk-content-schemas. This is because we are merging schemas into publishing api.
+
 # 4.4.1
 
 * Fix `Validator` module to handle JSON or other object types being passed as the payload ([#68](https://github.com/alphagov/govuk_schemas/pull/68))