+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # UNRELEASED
 
 * Update default content schemas url to point to publishing api rather than govuk-content-schemas. This is because we are merging schemas into publishing api.
+* Update path of allowed_document_types.yml to reflect new location in publishing api, allowing us to remove a symlink.
 
 # 4.4.1