+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update default content schemas url to point to publishing api rather than govuk-content-schemas. This is because we are merging schemas into publishing api.
 * Update path of allowed_document_types.yml to reflect new location in publishing api, allowing us to remove a symlink.
+* Introduce a setter method for manually configuring the path to schemas, outside of an env variable
 
 # 4.4.1