+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# UNRELEASED
+
+* Update default content schemas url to point to publishing api rather than govuk-content-schemas. This is because we are merging schemas into publishing api.
+* Update path of allowed_document_types.yml to reflect new location in publishing api, allowing us to remove a symlink.
+* Introduce a setter method for manually configuring the path to schemas, outside of an env variable
+
 # 4.4.1
 
 * Fix `Validator` module to handle JSON or other object types being passed as the payload ([#68](https://github.com/alphagov/govuk_schemas/pull/68))